requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
cffi==1.17.1
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
setuptools==75.3.0
sqlparse==0.5.1
